QOL changes to link sharing
