+++ frontend/src/components/Protectedroute.tsx
@@ -8,7 +8,7 @@ const Protectedroute = ({children}) => {
         return children;
     }
     else{
-        return <Navigate to="/login" replace/>
+        return <Navigate to="/login" state={{ from: true }} />;
     }
 }
